refactor to common base class
