+++ sktime/base/__init__.py
@@ -6,10 +6,12 @@ __author__ = ["mloning", "RNKuhns", "fkiraly"]
 __all__ = [
     "BaseObject",
     "BaseEstimator",
+    "BasePanelMixin",
     "_HeterogenousMetaEstimator",
     "load",
 ]
 
 from sktime.base._base import BaseEstimator, BaseObject
+from sktime.base._base_panel import BasePanelMixin
 from sktime.base._meta import _HeterogenousMetaEstimator
 from sktime.base._serialize import load
